requirements.txt
netCDF4==1.6.2
netCDF4==1.7.2
numpy==2.1.3
Pillow==11.0.0
pytorch_lightning==2.4.0
PyYAML==6.0.2
PyYAML==6.0.2
torch==2.5.1
torchmetrics==1.4.0.post0
torchvision==0.20.1
